+++ BGL/test/BGL/test_cgal_bgl_named_params.cpp
@@ -72,7 +72,6 @@ void test(const NamedParameters& np)
   assert(get_param(np, CGAL::internal_np::number_of_points_on_edges).v == 31);
   assert(get_param(np, CGAL::internal_np::nb_points_per_area_unit).v == 32);
   assert(get_param(np, CGAL::internal_np::nb_points_per_distance_unit).v == 33);
-  assert(get_param(np, CGAL::internal_np::new_face_visitor).v == 42);
   assert(get_param(np, CGAL::internal_np::throw_on_self_intersection).v == 43);
   assert(get_param(np, CGAL::internal_np::clip_volumes).v == 44);
   assert(get_param(np, CGAL::internal_np::use_compact_clipper).v == 45);
@@ -140,7 +139,6 @@ void test(const NamedParameters& np)
   check_same_type<31>(get_param(np, CGAL::internal_np::number_of_points_on_edges));
   check_same_type<32>(get_param(np, CGAL::internal_np::nb_points_per_area_unit));
   check_same_type<33>(get_param(np, CGAL::internal_np::nb_points_per_distance_unit));
-  check_same_type<42>(get_param(np, CGAL::internal_np::new_face_visitor));
   check_same_type<43>(get_param(np, CGAL::internal_np::throw_on_self_intersection));
   check_same_type<44>(get_param(np, CGAL::internal_np::clip_volumes));
   check_same_type<45>(get_param(np, CGAL::internal_np::use_compact_clipper));
@@ -207,7 +205,6 @@ int main()
                          .weight_calculator(A<39>(39))
                          .preserve_genus(A<40>(40))
                          .verbosity_level(A<41>(41))
-                         .new_face_visitor(A<42>(42))
                          .throw_on_self_intersection(A<43>(43))
                          .clip_volumes(A<44>(44))
                          .use_compact_clipper(A<45>(45))